Reduced visual clutter and removed unused code
Rather than having a separate textblock pop up if the category has a path set it's now displayed in the save to TextBox watermark.
If the checkbox is marked it will allow the user to overwrite it.
Rather than 2 lines for the checkbox and the textbox for save to it's now one merged control.